third_party/WebKit/LayoutTests/animations/interpolation/border-image-slice-interpolation-stability.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
test(function() {
  var player = target.animate([
    {borderImageSlice: '50'},
    {borderImageSlice: '50'},
  ], {
    duration: 1,
    fill: 'forwards',
    easing: 'cubic-bezier(0, 1.5, 1, 1.5)',
  });
  player.pause();
  player.currentTime = 0.6345195996109396
  assert_equals(getComputedStyle(target).borderImageSlice, '50');
});
</script>
